Move helper functions to a seperate file
