+++ ics/cost_update_helpers.py
@@ -1,0 +1,223 @@
+from django.contrib.postgres.aggregates import ArrayAgg
+from django.db.models.functions import Coalesce
+
+from ics.models import *
+from django.db.models import F, Count, Sum
+
+
+# calculate costs for all the children of current task
+def update_children(new_unit_cost, prev_unit_cost, updated_task, tasks, descendant_ingredients):
+	# calculate difference of unit costs
+	unit_cost = new_unit_cost - prev_unit_cost
+	# iterate through each child of the updated task and propagate data
+	for child in tasks[updated_task]['children']:
+		if child is not None:
+			new_difference = update_cost(updated_task, child, unit_cost, tasks, descendant_ingredients)
+			# call rec_cost() to propagate values to children of child task if it is present in tasks list
+			if child in tasks:
+				visited = {}  # handles tasks with circular dependencies
+				rec_cost(child, new_difference, tasks, visited, descendant_ingredients)
+
+
+# Make-shift solution: we're unable to filter out rashed tasks in ArrayAggs, so we just have to skip em
+def task_is_trashed(task, maps_of_non_trashed_tasks):
+	for map in maps_of_non_trashed_tasks:
+		if task not in map:
+			return True
+	return False
+
+
+# updates cost and remaining worth of tasks and returns new_difference which will be passed to child task
+def update_cost(task, child, unit_cost, tasks, descendant_ingredients):
+	if task_is_trashed(task, [descendant_ingredients, tasks]) or task_is_trashed(child, [descendant_ingredients, tasks]):
+		return 0  # @Aditya, does zero work?
+	# find number of parent contributing same ingredient to the child task
+	num_parents = len(descendant_ingredients[child]['task_ing_map'][(tasks[task]['process_type'], tasks[task]['product_type'])]['parent_tasks'])
+	# total amount of ingredient associated with the child task
+	total_amount = descendant_ingredients[child]['task_ing_map'][(tasks[task]['process_type'], tasks[task]['product_type'])]['amount']
+	# actual amount to be transferred to the child
+	actual_amount = float(total_amount / num_parents)
+	new_difference = unit_cost * actual_amount
+	tasks[child]['cost'] = tasks[child]['cost'] or 0
+	tasks[child]['remaining_worth'] = tasks[child]['remaining_worth'] or 0
+	# updated total cost and remaining_worth associated with the child and parent
+	print "new difference"
+	print new_difference
+	tasks[child]['cost'] = float(tasks[child]['cost']) + new_difference
+	tasks[child]['remaining_worth'] = float(tasks[child]['remaining_worth']) + new_difference
+	tasks[task]['remaining_worth'] = float(tasks[task]['remaining_worth']) - new_difference
+	# update child costs
+	Task.objects.filter(pk=child).update(cost=tasks[child]['cost'], remaining_worth=tasks[child]['remaining_worth'])
+	# update parent task costs
+	Task.objects.filter(pk=task).update(remaining_worth=tasks[task]['remaining_worth'])
+	return new_difference
+
+
+# function to recursively propagate data
+def rec_cost(parent, proportional_cost, tasks, visited, descendant_ingredients):
+	# batch_size should not be 0 (when you run script again)
+	unit_cost = float(round(proportional_cost / float(tasks[parent]['batch_size']), 2))
+	for child in tasks[parent]['children']:
+		if child is not None and child not in visited:
+			visited[child] = child
+			new_difference = update_cost(parent, child, unit_cost, tasks, descendant_ingredients)
+			if child in tasks:
+				rec_cost(child, new_difference, tasks, visited, descendant_ingredients)
+
+
+# fetch parents and children for list of related tasks
+def task_details(updated_task_descendants):
+	# contains task ids of descendants and ids of parents of descendants
+	related_tasks = set()
+	for task in list(updated_task_descendants):
+		related_tasks.add(task.id)
+	# for each descendant get parent id
+	parents = Task.objects.filter(is_trashed=False, pk__in=related_tasks).annotate(
+		task_parent_ids=ArrayAgg('inputs__input_item__creating_task__id'))
+
+	for parent in parents:
+		related_tasks = related_tasks.union(parent.task_parent_ids)
+
+	related_tasks_batch_size = Task.objects.filter(is_trashed=False, pk__in=related_tasks).annotate(
+		batch_size=Coalesce(Sum('items__amount'), 0))
+	related_tasks_with_parents_children = Task.objects.filter(is_trashed=False, pk__in=related_tasks).annotate(
+		children_list=ArrayAgg('items__inputs__task'),
+		task_parent_ids=ArrayAgg('inputs__input_item__creating_task__id'),
+		ingredients=ArrayAgg('task_ingredients__ingredient'))
+	tasks = {}
+	for x in range(0, related_tasks_with_parents_children.count()):
+		tasks[related_tasks_with_parents_children[x].id] = {'children': set(related_tasks_with_parents_children[x].children_list),
+						'process_type': related_tasks_batch_size[x].process_type_id,
+						'ingredients': set(related_tasks_with_parents_children[x].ingredients),
+						'product_type': related_tasks_batch_size[x].product_type_id, 'cost': related_tasks_batch_size[x].cost,
+					    'parents': set(related_tasks_with_parents_children[x].task_parent_ids),
+				    	'remaining_worth': related_tasks_batch_size[x].remaining_worth, 'batch_size': related_tasks_batch_size[x].batch_size}
+	return tasks
+
+
+# returns ingredient used by each descendant
+def descendant_ingredient_details(updated_task_descendants, tasks):
+	descendant_ingredients = {}
+	# populate task_ing_map with taskingredients along with all the parents contributing that ingredient for each descendant
+	for task in updated_task_descendants:
+		key = task.id
+		descendant_ingredients[key] = {}
+		task_ing_map = {}
+		for task_ing in tasks[key]['ingredients']:
+			ingredient_details = TaskIngredient.objects.filter(task_id=key, ingredient_id=task_ing).values('actual_amount',
+																						'ingredient__process_type', 'ingredient__product_type')[0]
+			task_ing_map[(ingredient_details['ingredient__process_type'], ingredient_details['ingredient__product_type'])] = {
+								'amount': ingredient_details['actual_amount'], 'parent_tasks': set()}
+		for parent in tasks[key]['parents']:
+			if task_is_trashed(parent, [tasks]):
+				continue
+			task_ing_map[(tasks[parent]['process_type'], tasks[parent]['product_type'])]['parent_tasks'].add(
+				parent)
+		descendant_ingredients[key]['task_ing_map'] = task_ing_map
+	return descendant_ingredients
+
+
+def get_non_trashed_descendants(task):
+	return Task.descendants(task).filter(is_trashed=False)
+
+
+# DELETED TASK HELPERS:
+
+def update_parents_of_deleted_task(updated_task_id):
+	# get all the ids of ingredients used and parents who contribute those ingredients
+	qs = Task.objects.filter(is_trashed=False, pk=updated_task_id).annotate(
+		task_parent_ids=ArrayAgg('inputs__input_item__creating_task__id'),
+		ingredients=ArrayAgg('task_ingredients__ingredient')
+	)
+	if qs.count() != 1 or qs[0].ingredients == [None]:
+		return
+	updated_task = qs[0]
+	parent_ids = updated_task.task_parent_ids
+	updated_task_parents = Task.objects.filter(pk__in=set(parent_ids)).annotate(
+		batch_size=Coalesce(Sum('items__amount'), 0))
+
+	# map ingredients with parents
+	ingredients_to_parents_map = {}
+	print(updated_task.ingredients)
+	for ingredient_id in updated_task.ingredients:
+		ingredient = Ingredient.objects.get(pk=ingredient_id)
+		parents_contributing_ingredient = {}
+		for x in range(0, len(set(parent_ids))):
+			if updated_task_parents[x].process_type_id == ingredient.process_type_id and \
+							updated_task_parents[x].product_type_id == ingredient.product_type_id:
+				parents_contributing_ingredient[updated_task_parents[x].id] = updated_task_parents[x]
+		ingredients_to_parents_map[ingredient.id] = parents_contributing_ingredient
+
+	# iterate over each ingredient
+	print(ingredients_to_parents_map)
+	for ingredient_id, parents_contributing_ingredient in ingredients_to_parents_map.iteritems():
+		old_amount = TaskIngredient.objects.filter(ingredient_id=ingredient_id).values('actual_amount')[0]['actual_amount']
+		new_amount = 0
+		num_parents = len(parents_contributing_ingredient)
+		update_parents_for_ingredient(parents_contributing_ingredient, old_amount, new_amount, num_parents)
+
+
+def update_parents_for_ingredient(parents_contributing_ingredient, old_amount, new_amount, num_parents, creating_task=-1,
+								  input_added=False, input_deleted=False):
+	old_avg_amount = old_amount / num_parents
+	new_avg_amount = new_amount / num_parents
+	total_diff = 0
+	print parents_contributing_ingredient
+	for task in parents_contributing_ingredient:
+		if parents_contributing_ingredient[task].cost is None:
+			return
+		print "parent cost"
+		print parents_contributing_ingredient[task].cost
+		print "parent batch size"
+		print parents_contributing_ingredient[task].batch_size
+		unit_cost = parents_contributing_ingredient[task].cost / parents_contributing_ingredient[task].batch_size
+		# cost of ingredient used previously
+		prev_utilization = old_avg_amount * unit_cost
+		if task == creating_task and input_added:
+			prev_utilization = 0
+		# cost of ingredient used now
+		new_utilization = new_avg_amount * unit_cost
+		if task == creating_task and input_deleted:
+			new_utilization = 0
+		print "prev util"
+		print prev_utilization
+		print "new util"
+		print new_utilization
+		utilization_diff = new_utilization - prev_utilization
+		print "utilization diff"
+		print utilization_diff
+		if utilization_diff > parents_contributing_ingredient[task].remaining_worth:
+			print "if"
+			new_remaining_worth = 0
+			total_diff += parents_contributing_ingredient[task].remaining_worth
+		elif parents_contributing_ingredient[task].remaining_worth - utilization_diff < parents_contributing_ingredient[task].cost:
+			print "elif"
+			new_remaining_worth = parents_contributing_ingredient[task].remaining_worth - utilization_diff
+			total_diff += utilization_diff
+		else:
+			print "else"
+			new_remaining_worth = parents_contributing_ingredient[task].cost
+			total_diff += parents_contributing_ingredient[task].remaining_worth - parents_contributing_ingredient[task].cost
+		# Task.objects.filter(pk=task).update(remaining_worth=new_remaining_worth)
+		print "new remaining"
+		print new_remaining_worth
+		print "total diff"
+		print total_diff
+	return total_diff
+
+
+def update_children_of_deleted_task(deleted_task):
+	# fetch all the descendants of deleted task
+	deleted_task_descendants = get_non_trashed_descendants(Task.objects.filter(pk=deleted_task)[0])
+	if not deleted_task_descendants:
+		print ("No descendents")
+		return
+	tasks = task_details(deleted_task_descendants)
+	descendant_ingredients = descendant_ingredient_details(deleted_task_descendants, tasks)
+	if tasks[deleted_task]['children'] != set([None]):
+		batch_size = tasks[deleted_task]['batch_size']
+		cost = tasks[deleted_task]['cost']
+		prev_unit_cost = float(round(cost / batch_size, 2))
+		new_unit_cost = 0
+		# print("cost: %d, batch_size: %d, prev_unit_cost: %d, new_unit_cost: %d" % (cost, batch_size, prev_unit_cost, new_unit_cost))
+		update_children(new_unit_cost, prev_unit_cost, deleted_task, tasks, descendant_ingredients)
